competitor_intel_pro/frontend/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
pyyaml==6.0.2
openai==0.28.1
playwright==1.45.0
pillow==10.4.0
streamlit==1.38.0
matplotlib==3.9.2
reportlab==4.2.5